+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.13] / 2023-06-30
+### Features
+- Add Aps Loggin and logger test count
+### Application
+- Add `ApsApplicationLogger`
+- Add `ricaun.Auth.Aps.UI` version `1.0.1`
+### Console
+- Add time delay close and Kill process
+### TestAdapter
+- Add `OrdinalIgnoreCase` to fix Metadata Mapper
+
 ## [1.0.12] / 2023-06-26
 ### Features
 - Add authentication with `ricaun.Auth`
@@ -170,6 +181,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [x] TestsFail
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.0.13]: ../../compare/1.0.12...1.0.13
 [1.0.12]: ../../compare/1.0.11...1.0.12
 [1.0.11]: ../../compare/1.0.10...1.0.11
 [1.0.10]: ../../compare/1.0.9...1.0.10